final commit

+++ database/migrations/2024_03_19_000002_create_centroids_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->float('tanggungan_num');
             $table->enum('kondisi_rumah', ['baik', 'cukup', 'kurang']);
             $table->enum('status_kepemilikan', ['hak milik', 'numpang', 'sewa']);
-            $table->float('penghasilan_num');
+            $table->bigInteger('penghasilan_num');
             $table->integer('tahun');
             $table->integer('periode');
             $table->string('keterangan')->nullable();
